requirements.txt
python-jenkins
Django>=1.10.4
